+++ src/extra.ts
@@ -1,4 +1,8 @@
-import { CreateNeuronRequest } from "./types/intercanister";
+// API calls with a scope wider than one canister:
+// - calls that require multiple canisters.
+// - calls that require additional systems such as hardware wallets.
+
+import { CreateNeuronRequest } from "./types/extra";
 import { Principal } from "@dfinity/principal";
 import { LedgerCanister } from "./ledger";
 import { NnsDappCanister } from "./nns_dapp";
